fix: 增加文档,调整一些下载问题

+++ src/lib/download.js
@@ -13,35 +13,26 @@ const zipFilenameDic = {
   react: 'vue3-vite-template.zip',
 };
 
-async function dlTemplate(tplType) {
+async function dlTemplate(tplType, onlyDownload = false) {
+  // 读取配置文件
   const exist = await fse.pathExists(cfgPath);
-  if (exist) {
-    await dlAction(tplType);
-  } else {
+  if (!exist) {
     await defConfig();
-    await dlAction(tplType);
   }
-}
 
-async function dlAction(tplType) {
-  try {
-    const tplPath = tools.getTplPath(tplType);
-    await fse.remove(tplPath);
-    await fse.ensureDir(tplPath);
-  } catch (err) {
-    console.err(err);
-    process.exit();
-  }
+  // 确保下载路径
+  const tplPath = tools.getTplPath(onlyDownload);
+  await fse.remove(tplPath);
+  await fse.ensureDir(tplPath);
 
   const jsonConfig = await fse.readJson(cfgPath);
-  const dlSpinner = ora(chalk.cyan('Downloading template...'));
+  const dlSpinner = ora(chalk.cyan('正在下载模板...'));
   dlSpinner.start();
 
-  // 声明路径
+  // 声明路径变量
   const zipFilename = zipFilenameDic[tplType];
   const downloadUrl = jsonConfig.mirror + zipFilename;
-  const targetDir = path.resolve(__dirname, 'template');
-  const targetFile = path.resolve(targetDir, zipFilename);
+  const targetFile = path.resolve(tplPath, zipFilename);
 
   try {
     const res = await fetch(downloadUrl);
@@ -55,17 +46,18 @@ async function dlAction(tplType) {
         resolve();
       });
     });
-    await extract(targetFile, { dir: targetDir });
-    await fse.remove(targetFile);
+    // 如果不是仅下载模板,就解压缩并删除zip文件
+    if (!onlyDownload) {
+      await extract(targetFile, { dir: tplPath });
+      await fse.remove(targetFile);
+    }
   } catch (err) {
-    dlSpinner.text = chalk.red(
-      `Download template(${downloadUrl}) failed. ${err}`
-    );
+    dlSpinner.text = chalk.red(`下载模板(${downloadUrl}) 失败 ${err}`);
     dlSpinner.fail();
     process.exit();
   }
 
-  dlSpinner.text = 'Download template successful.';
+  dlSpinner.text = '下载模板成功.';
   dlSpinner.succeed();
 }
